Look for markdown preview in all panes to update
Previously only the next pane was checked to see if a preview
exists to update.  This would fail to update the preview if
it was moved moved from its original pane to a different pane.

Now the view tracks the buffer and updates on 'saved' events and
shows the preview if it is not in the active pane.