Add platform-specific key bindings
It is preferable to offer key bindings that correspond to the modifier keys used by the different operating systems on which Atom functions. In this case, only a Windows key binding was provided. This commit changes this by specifying that the previous key binding only applies to Windows and by adding a key binding for Mac OS X that uses the command key.