Add ability to use HTML tags in Markdown documents
The original Markdown specification by John Gruber suggests using plain HTML for
tables and other advances things. GFM doesn't document the use of HTML but
it is supported and correctly rendered everywhere on GitHub (Issues, README.md,
etc.)

Because of this, `markdown-preview` should support the use of HTML too,
otherwise it wouldn't be a Markdown preview tool.