Add configurable pane direction.
Addresses #99 and part of #328.  Does _not_ address any kind of "opposite" setting.

Migrates existing true/false (pane to the right or not) config settings into "right"/"none" equivalent value.

Note: In order to migrate an exiting setting, we have to run code very early when the module loads, _before_ it activates.  This is unfortunate from a performance standpoint.  If we're willing to simply lose any previous configuration setting (falling back to the default, "right"), we can avoid the perf impact of doing work during module load.  I'm completely open to the wisdom of the Atom community here; I just wanted to start with an already-backward-compatible option.