:bug: Remove expectation of `decimal` from generated HTML.
There are two tests that currently fail on master: `when core:save-as is triggered`
and `when core:copy is triggered`.  Neither of these seem to have _anything_ to do
with this package; there's a `decimal` class that used to be expected in the
generated HTML output that isn't there any longer.