Make markdown pane element scroll to editor's position
On any change in markdown text, the markdown preview pane re-renders, losing the scroll position in process
This commit attempts to fix that by making markdown preview pane scroll to position of markdown editor.

This commit does not take into account thing like images, etc in markdown that take up more vertical space but
it is simple and will work well for markdown without elements that expand to be to big.